requirements.txt
fastapi
uvicorn
python-dotenv

streamlit
requests

langchain
langgraph
langchain-community
langsmith

google-generativeai
supabase
